+++ changelog.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v9.0.1...HEAD)
+### Added
+* [PCM-2302](https://inindca.atlassian.net/browse/PCM-2302) Added queue for headset events to prevent issues brought on by race conditions between events
 ### Changed
 * [PCM-2297](https://inindca.atlassian.net/browse/PCM-2297) Update genesys-cloud-client-logger and genesys-cloud-streaming-client to fix Snyk vulnerabilities.
 ### Fixed